+++ OW_dev_sn.h
@@ -26,17 +26,6 @@
 	name[idx].sn.serial_number = OWGetSerialNumber(pROM);				//!< Macro to set working defaults for peripheral \b name on index \b idx
 
 
-#define OW_SN_GETTER(name)														\
-/*!\brief name Serial Number getter												\
-** \param[in] pCpnt - Pointer to name peripheral								\
-** \return name peripheral serial number										\
-**/																				\
-__INLINE uint64_t NONNULL_INLINE__ name##_SN_Get(name##_t * const pCpnt) {		\
-	uint64_t SN;																\
-	OW_SN_Get(&pCpnt->sn, &SN);													\
-	return SN; }														//!< Macro to generate serial number getter for peripheral \b name
-
-
 #define OW_SN_OFFSET(name)	OW_PERIPHERAL_DEV_OFFSET(name, sn)			//!< Macro to get sn structure offset in \b name peripheral structure
 
 
@@ -58,9 +47,10 @@ typedef struct {
 
 /*!\brief OneWire serial number getter for sn type device
 ** \param[in] pSN - Pointer to serial number device type structure
-** \param[in,out] pSerialNumber - Pointer to output serial number
+** \return Serial Number
 **/
-void NONNULL__ OW_SN_Get(const OW_sn_t * const pSN, uint64_t * const pSerialNumber);
+__INLINE uint64_t NONNULL_INLINE__ OW_SN_Get_SN(const OW_sn_t * const pSN) {
+	return pSN->serial_number; }
 
 
 /****************************************************************/